+++ src/common/ClickCardBox.jsx
@@ -1,8 +1,8 @@
 import { useContext, useCallback } from 'react';
 
-import * as S from 'styles.js';
 import { CardDispatchContext } from 'store/card/CardContext';
 import { TYPES } from 'store/card/types';
+import PointerBox from 'components/PointerBox/PointerBox';
 
 export default function ClickCardBox({ children }) {
   const dispatch = useContext(CardDispatchContext);
@@ -11,5 +11,5 @@ export default function ClickCardBox({ children }) {
     dispatch({ type: TYPES.SET_LIST_MODAL_FLAG, flag: true });
   }, []);
 
-  return <S.PointerBox onClick={onClickCard}>{children}</S.PointerBox>;
+  return <PointerBox onClick={onClickCard}>{children}</PointerBox>;
 }
