--- src/components/AnotherCard/AnotherCard.jsx
@@ -1,11 +1,0 @@
-import * as S from 'styles.js';
-
-export default function AnotherCard({ onClick }) {
-  return (
-    <S.CardAlignBox onClick={onClick}>
-      <S.Card isSmall={true} color={'#e5e5e5'}>
-        <S.Sign>{'+'}</S.Sign>
-      </S.Card>
-    </S.CardAlignBox>
-  );
-}
