+++ src/components/Form/index.jsx
@@ -1,3 +1,3 @@
-export default function FormBasic({ onSubmitForm, payload, children }) {
+export default function Form({ onSubmitForm, payload, children }) {
   return <form onSubmit={(event) => onSubmitForm(event, payload)}>{children}</form>;
 }
