+++ src/containers/CardOwnerInput/CardOwner.stories.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import CardOwner from 'containers/CardOwnerInput.jsx/CardOwner';
+import CardOwner from 'containers/CardOwnerInput/CardOwner';
 
 export default {
   title: 'CardAddition/CardOwner',
