+++ src/index.jsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom/client';
 import App from 'App';
-import * as S from 'styles.js';
+import * as S from 'globalStyles/index.js';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
